+++ piwigo/Album/Discover/TaggedImagesViewController.m
@@ -1704,7 +1704,7 @@
                 }
             });
             // Should we continue loading images?
-            NSLog(@"==> Should we continue loading images? (scrolled to %ld)", self.didScrollToImageIndex);
+            NSLog(@"==> Should we continue loading images? (scrolled to %ld)", (long)self.didScrollToImageIndex);
             if (self.didScrollToImageIndex >= newDownloadedImageCount) {
                 [self needToLoadMoreImages];
             }
